BinaryStream: generalize the API and make it work with const references
The function ```c++
void f(BinaryStream& bs, const std::vector<int>& v) {
  bs << v;
}
```
failed to compile because
`template <typename T> bool BinaryStream::operator<<(const& T)`
was used instead of
`template <typename T> bool BinaryStream::operator<<(std::vector<T>&)`.
In this patch, BinaryStream API was redesigned, duplicating code was
removed and `BinaryStream::operator<<` was made less fragile when
working with const references. The changes are fully backwards
compatible.

Note that normally there is no need for serialization to mutate the
object, so `BinaryStream::operator<<` should be working with const
references only. However, `SerialisableObject::SerialiseWrapper` is not
a constant method because it does both serialization and
deserialization. A proper design would be to separate this logic to two
methods. This is to be considered in the future. For now, since
instances of class SerialisableObject can appear in vectors and other
containers, we have to require non-const references for the
corresponding specializations of `BinaryStream::opreator<<`.