BinaryStream::(de-)serialise(std::string& rhs): revert to using &rhs[0] instead of rhs.data()
In C++ before C++11, meddling with a string using the pointer returned
by `string.data()`, `string.c_str()` or `&string[0]` are undefined
behaviour.  `std::string` was not required storing its contents in a
continuous buffer, and GCC implemented copy-on-write strings.
Nevertheless, this code apparently worked in the environments it was
used, because GCC does store string contents in a continuous buffer, and
calling `resize()` before deserializing probably helped with
copy-on-writing. In newer standards of C++, this is no longer an
undefined behaviour.