BStore: implement JSON encoding
Implement BStore::JsonEncode methods that can be used to serialise
BStore contents in JSON format, if this representation is possible.
This requires type checking enabled. BStore::operator>> is not changed
to provide backwards compatibility and a way to serialise a BStore with
type checking disabled (only useful if all values are strings).

Functions in Json.h can also be used independently to serialise other
C++ integral types, strings, vectors and maps.