Store: store structured data as JSON objects
* Use JSON encoding for strings with special characters, vectors, and
  maps to ensure that deserialization returns what has been serialized.

* Make JsonParser work for JSON objects containing strings with special
  characters.

* JsonParser now returns a bool indicating whether the parsing was
  successful.

* Replace `template <typename T> void operator>>(T&)` with `void
  operator>>(std::string&)` since it only worked for strings anyways.

* Make `operator>>` produce numbers as JSON numbers.

* Rewrite UnitTests/StoreTest to thoroughly test Store implementation.

The type of the stored object is encoded by the first character in the
stored value: `"` for strings, `[` for arrays, and `{` for objects. A
matching character is expected at the end of the value. Any other
character at the beginning of the value means that the object was
encoded with a custom `operator<<(std::ostream&, T)`, and it is expected
that `operator>>(std::istream&, T&)` can be used to decode the object
completely. If such encoding produces a string beginning with one of the
characters above, it is wrapped into a pair of '"'. These conversions
are handled transparently to the user.