Store::Initialise: fix some edge cases and make consistent with the new store contents
Calling Initialise on a file setting an integer variable, e.g.,

    var 1

resulted in this variable stored as "1" and interpreted as a string, so
Get<int>("1") returned 0 (json_decode failed to parse "1" as a number).
This is inconsistent both with the previous and the new behaviours where
numbers set with `Set` are stored unquoted. This patch updates
Initialise to store a value as is, unless it begins with a `"`, or it
begins with a `[` or `{` and is not a valid JSON object. In these cases
the value is quoted.

Also:

* Initialise now does not drop the whitespace in the value, so

    var a string with a   tab and a   long space

does not result in `a string with a tab and a long space`.

* Comments, newlines, and spaces at the end can now be escaped:

    escaped_comment test \# test
    escaped_newline test \
    test2
    escaped_whitespace test \

* One backward incompatible change is that a space is now not required
  before an inline comment:

    var val#ue

now results in `val` rather than `val#ue`.

StoreTest has been updated to test for these cases.